Fix removing project errors, add task details
When removing the projects, the tasks still wanted to display, but the project data was gone, so an error was displayed. Also when all the projects were removed, there was another error, so now a check was added to see if there are no projects and if there are none, text is written that showcases that.

Task details can be shown by clicking on a specific task. There is also a new edit button, that allows the user to change task data.